Add dish dto, refactor dish contollers

+++ src/main/java/pl/nojman/restaurant_api/Controllers/DishController.java
@@ -3,6 +3,8 @@ package pl.nojman.restaurant_api.Controllers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.nojman.restaurant_api.Dtos.DishDto;
+import pl.nojman.restaurant_api.Mappers.Mapper;
 import pl.nojman.restaurant_api.Models.Dish;
 import pl.nojman.restaurant_api.Models.Restaurant;
 import pl.nojman.restaurant_api.Services.DishService;
@@ -10,6 +12,7 @@ import pl.nojman.restaurant_api.Services.RestaurantService;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/restaurants/{restaurantId}/dishes")
@@ -17,40 +20,64 @@ public class DishController {
 
     private DishService service;
     private RestaurantService restaurantService;
+    private Mapper mapper;
 
     @Autowired
-    public DishController(DishService service, RestaurantService restaurantService) {
+    public DishController(DishService service, RestaurantService restaurantService, Mapper mapper) {
         this.service = service;
         this.restaurantService = restaurantService;
+        this.mapper = mapper;
     }
 
     @GetMapping()
-    public List<Dish> getDishes(@PathVariable long restaurantId) {
+    public ResponseEntity<List<DishDto>> getDishes(@PathVariable long restaurantId) {
         Optional<Restaurant> restaurant = restaurantService.getRestaurant(restaurantId);
         if (restaurant.isPresent()) {
-            return this.service.findAll(restaurantId);
+            return ResponseEntity.ok(this.service.findAll(restaurantId).stream()
+                    .map(x -> this.mapper.dishModelToDto(x))
+                    .collect(Collectors.toList()));
         }
-        return null;
+        return ResponseEntity.notFound().build();
     }
 
     @GetMapping("/{dishId}")
-    public Dish getDish(@PathVariable long restaurantId, @PathVariable long dishId) {
-        return this.service.find(restaurantId, dishId).get();
+    public ResponseEntity<DishDto> getDish(@PathVariable long restaurantId, @PathVariable long dishId) {
+        Optional<Dish> dish = this.service.find(restaurantId, dishId);
+        if (dish.isPresent()) {
+            return ResponseEntity.ok(this.mapper.dishModelToDto(dish.get()));
+        }
+        return ResponseEntity.notFound().build();
     }
 
-    @PostMapping()
-    public void create(@PathVariable long restaurantId, @RequestBody Dish dish) {
-        this.service.create(dish);
-    }
+//    @PostMapping()
+//    public ResponseEntity<Void> create(@PathVariable long restaurantId, @RequestBody DishDto dishDto) {
+//        this.service.create(dish);
+//    }
 
     @DeleteMapping("/{dishId}")
-    public void delete(@PathVariable long restaurantId, @PathVariable long dishId) {
-        this.service.delete(dishId);
+    public ResponseEntity<Void> delete(@PathVariable long restaurantId, @PathVariable long dishId) {
+        Optional<Restaurant> restaurant = this.restaurantService.getRestaurant(restaurantId);
+        if (restaurant.isPresent()) {
+            Optional<Dish> dish = this.service.find(restaurantId, dishId);
+            if (dish.isPresent()) {
+                this.service.delete(dishId);
+                return ResponseEntity.accepted().build();
+            }
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{dishId}")
-    public void update(@PathVariable long restaurantId, @PathVariable long dishId, @RequestBody Dish dish) {
-        this.service.update(dish);
+    public ResponseEntity<Void> update(@PathVariable long restaurantId, @PathVariable long dishId, @RequestBody DishDto dishDto) {
+        Optional<Restaurant> restaurant = this.restaurantService.getRestaurant(restaurantId);
+        if (restaurant.isPresent()) {
+            Optional<Dish> dish = this.service.find(restaurantId, dishId);
+            if (dish.isPresent()) {
+                this.service.update(dishDto);
+                return ResponseEntity.accepted().build();
+            }
+        }
+        return ResponseEntity.notFound().build();
     }
 
 }
